Move package under src and update paths
