+++ src/vectordb/__init__.py
@@ -1,0 +1,6 @@
+"""Simple in-memory vector database package."""
+
+from .db import VectorDB, INDEX_PATH, DATA_PATH, MODEL_NAME
+from .api import create_app
+
+__all__ = ["VectorDB", "create_app", "INDEX_PATH", "DATA_PATH", "MODEL_NAME"]
